requirements.txt
# Stock Risk Modelling App - Enterprise Edition v4.0
# ==================================================
# pip install -r requirements.txt

# Core Framework
streamlit>=1.35.0

# Data Sources
yfinance>=0.2.40
pandas-datareader>=0.10.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0

# Statistical & Risk Models
scipy>=1.11.0
arch>=6.2.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0

# PDF Report Generation
fpdf2>=2.7.0

# Static Image Export for Charts (optional - for PDF charts)
# kaleido>=0.2.1

# Environment Configuration
python-dotenv>=1.0.0
